+++ CHANGES.rst
@@ -9,6 +9,7 @@ Here you can see the full list of changes between each SQLAlchemy-Continuum rele
 
 - Added optional native trigger based versioning for PostgreSQL dialect
 - Added create_models option
+- Added count_versions utility function
 - Fixed custom transaction column name handling with models using joined table inheritance
 - Fixed subquery strategy support for models using joined table inheritance