+++ CHANGES.rst
@@ -12,6 +12,7 @@ Here you can see the full list of changes between each SQLAlchemy-Continuum rele
 - Added count_versions utility function
 - Fixed custom transaction column name handling with models using joined table inheritance
 - Fixed subquery strategy support for models using joined table inheritance
+- Fixed savepoint handling
 
 
 1.0.3 (2014-07-16)