+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 Here you can see the full list of changes between each SQLAlchemy-Continuum release.
 
 
-1.1.0 (2014-xx-xx)
+1.1.0 (2014-09-02)
 ^^^^^^^^^^^^^^^^^^
 
 - Added optional native trigger based versioning for PostgreSQL dialect